+++ Source/Alternet.UI.Pal/Api/PageSetupDialog.Api.h
@@ -43,10 +43,10 @@ ALTERNET_UI_API void PageSetupDialog_SetDocument_(PageSetupDialog* obj, PrintDoc
     #endif
 }
 
-ALTERNET_UI_API double PageSetupDialog_GetMinMarginLeft_(PageSetupDialog* obj)
+ALTERNET_UI_API float PageSetupDialog_GetMinMarginLeft_(PageSetupDialog* obj)
 {
     #if !defined(__WXMSW__) || defined(_DEBUG)
-    return MarshalExceptions<double>([&](){
+    return MarshalExceptions<float>([&](){
     #endif
         return obj->GetMinMarginLeft();
     #if !defined(__WXMSW__) || defined(_DEBUG)
@@ -54,7 +54,7 @@ ALTERNET_UI_API double PageSetupDialog_GetMinMarginLeft_(PageSetupDialog* obj)
     #endif
 }
 
-ALTERNET_UI_API void PageSetupDialog_SetMinMarginLeft_(PageSetupDialog* obj, double value)
+ALTERNET_UI_API void PageSetupDialog_SetMinMarginLeft_(PageSetupDialog* obj, float value)
 {
     #if !defined(__WXMSW__) || defined(_DEBUG)
     MarshalExceptions<void>([&](){
@@ -65,10 +65,10 @@ ALTERNET_UI_API void PageSetupDialog_SetMinMarginLeft_(PageSetupDialog* obj, dou
     #endif
 }
 
-ALTERNET_UI_API double PageSetupDialog_GetMinMarginTop_(PageSetupDialog* obj)
+ALTERNET_UI_API float PageSetupDialog_GetMinMarginTop_(PageSetupDialog* obj)
 {
     #if !defined(__WXMSW__) || defined(_DEBUG)
-    return MarshalExceptions<double>([&](){
+    return MarshalExceptions<float>([&](){
     #endif
         return obj->GetMinMarginTop();
     #if !defined(__WXMSW__) || defined(_DEBUG)
@@ -76,7 +76,7 @@ ALTERNET_UI_API double PageSetupDialog_GetMinMarginTop_(PageSetupDialog* obj)
     #endif
 }
 
-ALTERNET_UI_API void PageSetupDialog_SetMinMarginTop_(PageSetupDialog* obj, double value)
+ALTERNET_UI_API void PageSetupDialog_SetMinMarginTop_(PageSetupDialog* obj, float value)
 {
     #if !defined(__WXMSW__) || defined(_DEBUG)
     MarshalExceptions<void>([&](){
@@ -87,10 +87,10 @@ ALTERNET_UI_API void PageSetupDialog_SetMinMarginTop_(PageSetupDialog* obj, doub
     #endif
 }
 
-ALTERNET_UI_API double PageSetupDialog_GetMinMarginRight_(PageSetupDialog* obj)
+ALTERNET_UI_API float PageSetupDialog_GetMinMarginRight_(PageSetupDialog* obj)
 {
     #if !defined(__WXMSW__) || defined(_DEBUG)
-    return MarshalExceptions<double>([&](){
+    return MarshalExceptions<float>([&](){
     #endif
         return obj->GetMinMarginRight();
     #if !defined(__WXMSW__) || defined(_DEBUG)
@@ -98,7 +98,7 @@ ALTERNET_UI_API double PageSetupDialog_GetMinMarginRight_(PageSetupDialog* obj)
     #endif
 }
 
-ALTERNET_UI_API void PageSetupDialog_SetMinMarginRight_(PageSetupDialog* obj, double value)
+ALTERNET_UI_API void PageSetupDialog_SetMinMarginRight_(PageSetupDialog* obj, float value)
 {
     #if !defined(__WXMSW__) || defined(_DEBUG)
     MarshalExceptions<void>([&](){
@@ -109,10 +109,10 @@ ALTERNET_UI_API void PageSetupDialog_SetMinMarginRight_(PageSetupDialog* obj, do
     #endif
 }
 
-ALTERNET_UI_API double PageSetupDialog_GetMinMarginBottom_(PageSetupDialog* obj)
+ALTERNET_UI_API float PageSetupDialog_GetMinMarginBottom_(PageSetupDialog* obj)
 {
     #if !defined(__WXMSW__) || defined(_DEBUG)
-    return MarshalExceptions<double>([&](){
+    return MarshalExceptions<float>([&](){
     #endif
         return obj->GetMinMarginBottom();
     #if !defined(__WXMSW__) || defined(_DEBUG)
@@ -120,7 +120,7 @@ ALTERNET_UI_API double PageSetupDialog_GetMinMarginBottom_(PageSetupDialog* obj)
     #endif
 }
 
-ALTERNET_UI_API void PageSetupDialog_SetMinMarginBottom_(PageSetupDialog* obj, double value)
+ALTERNET_UI_API void PageSetupDialog_SetMinMarginBottom_(PageSetupDialog* obj, float value)
 {
     #if !defined(__WXMSW__) || defined(_DEBUG)
     MarshalExceptions<void>([&](){
